+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.0 (unreleased)
 
+- Added `AnomalyDetector` struct
 - Changed `anomalies` method to return slice
 
 ## 0.2.3 (2022-03-19)